Migrated to /boot API call
In order to reduce the number of 'empty' calls to kernelci API, the
tasks were switched to use /boot API. The result is we only get the
successful boot reports from the boards of interest. The builds that do
not include any boots that are of interest, are not queried any more.
This means the number of calls to the API is only equal to:
 number of jobs * numer of boards
For each result there is an additional call to /build API in order to
retrieve the path to build artifacts.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>