tasks: removed hard coded query date_range
Since frequency of monitor_boots can be set from admin, this change
removes assumption that admin has to be in sync with settings. Last
check on job/board pair is saved to DB. After initial sync which is 24
hours (1440 min) the subsequent syncs depend on the check frequency.
Also time_range was used instead of date_range for better resolution.
Since minimal interval for kernelci is 10 min, shorter checks will be
ignored.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>